requirements.txt
tensorflow==2.4.0
lxml==4.5.2
validators==0.18.2
pandas==1.0.5
numpy==1.18.5
streamlit==0.79.0
typing_extensions==3.7.4.3
brotli==1.0.9
cryptography==3.4.7
Cython==0.29.22
docutils==0.16
importlib_metadata==3.10.0
keyring==23.0.1
lockfile==0.12.2
ordereddict==1.1
Pillow==8.1.2
protobuf==3.15.6
pyOpenSSL==20.0.1
wincertstore==0.2
zipp==3.4.1
